--- pass_project/app/templates/app/draft_pdf_template.html
@@ -1,9 +1,0 @@
-<!-- app/templates/app/draft_pdf_template.html -->
-<!DOCTYPE html>
-<html>
-<head><meta charset="UTF-8"></head>
-<body>
-  <h1>{{ draft.draft_name }}</h1>
-  <pre>{{ draft.create_draft }}</pre>
-</body>
-</html>
