--- pass_project/app/templatetags/form_tags.py
@@ -1,12 +1,0 @@
-# app/templatetags/form_tags.py
-from django import template
-register = template.Library()
-
-@register.filter(name='add_class')
-def add_class(field, css_class):
-    return field.as_widget(attrs={"class": css_class})
-
-@register.filter(name='attr')
-def set_attr(field, attr_value):
-    attribute, value = attr_value.split(':')
-    return field.as_widget(attrs={attribute: value})
