+++ pass_project/assist/static/assist/editor.css
