+++ pass_project/core/templates/core/main.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>메인 페이지</title>
+</head>
+<body>
+    <h1>메인 페이지</h1>
+    <p>환영합니다!</p>
+</body>
+</html>
