+++ pass_project/core/views.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def main_view(request):
+    return render(request, 'core/main.html')
