+++ pass_project/pass_project/urls.py
@@ -19,5 +19,8 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('template/', include('app.urls')),
+    path('', include('core.urls')),           # 메인 페이지
+    path('accounts/', include('accounts.urls')), # 로그인, 마이페이지
+    path('user_admin/', include('user_admin.urls')),    # 관리자 페이지
+    path('assist/', include('assist.urls')),  # AI 기능
 ]
