+++ pass_project/requirements.txt
@@ -4,3 +4,8 @@ xhtml2pdf==0.2.17
 python-docx==1.1.2
 requests==2.32.3
 python-dotenv==1.1.0
+transformers==4.40.0
+torch==2.2.0
+accelerate==0.28.0
+safetensors==0.4.2
+protobuf==3.20.0
