+++ pass_project/user_admin/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AdminConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "user_admin"
